+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 `Unreleased`_
 =============
 
+- |Feature| Allow savings metrics to be used in :class:`~empulse.models.CSBoostClassifier` and
+  :class:`~empulse.models.CSLogitClassifier` as the objective function.
 - |Fix| Fix error when importing Empulse without any optional dependencies installed.
 
 `0.8.0`_ (01-06-2025)